+++ deepinv/models/__init__.py
@@ -35,3 +35,4 @@ from .gan import (
 )
 from .complex import to_complex_denoiser
 from .dynamic import TimeAgnosticNet, TimeAveragingNet
+from .unet import UNet