+++ examples/sampling/demo_sde_image_generation.py
@@ -11,7 +11,7 @@ This method requires:
 * A well-trained denoiser with varying noise levels (ideally with large noise levels) (e.g., :class:`deepinv.model.edm.NCSNpp`).
 * Define a drift term :math:`f(x, t)` and a diffusion term :math:`g(t)` for the forward-time SDE.
 
-The forward-time SDE is defined as follows, for :math:`t \in \[0, T\]`:
+The forward-time SDE is defined as follows, for :math:`t \in [0, T]`:
 
 .. math::
     d\, x_t = f(x_t, t) d\,t + g(t) d\, w_t.
@@ -20,21 +20,21 @@ Let :math:`p_t` denote the distribution of the random vector :math:`x_t`.
 The reverse-time SDE is defined as follows, running backward in time:
 
 .. math::
-    d\, x_t = \(f(x_t, t) - g(t)^2 \nabla \log p_t(x_t) \) d\,t + g(t) d\, w_t.
+    d\, x_t =\left(f(x_t, t) - g(t)^2 \nabla \log p_t(x_t)\right( d\,t + g(t) d\, w_t.
 
 This reverse-time SDE can be used as a generative process. 
 The (Stein) score function :math:`\nabla p_t(x_t)` can be approximated by Tweedie's formula. In particular, if 
 
-..math::
-    x_t \vert x_0 \sim \mathcal{N}(\mu_tx_0, \sigma_t^2 \Id),
+.. math::
+    x_t \vert x_0 \sim \mathcal{N}(\mu_tx_0, \sigma_t^2 \mathrm{Id}),
 
 then
 
 .. math::
-    \nabla p_t(x_t) = \frac{\left(D_{\sigma_t}(x_t) - \mu_t x_t \right)}{\sigma_t^2}}.
+    \nabla p_t(x_t) = \frac{\mu_t  D_{\sigma_t}(x_t) -  x_t }{\sigma_t^2}.
 
 Starting from a random point following the end-point distribution :math:`p_T` of the forward process, 
-solving the reverse-time SDE gives us a sample of the data distribution :math:`p_0'.
+solving the reverse-time SDE gives us a sample of the data distribution :math:`p_0`.
 """
 
 # %% Define the forward process
@@ -63,8 +63,9 @@ denoiser = EDMPrecond(model=unet).to(device)
 
 # %%
 # Define the SDE. In this example, we use the Variance-Exploding SDE, whose forward process is defined as:
+#
 # .. math::
-#     d\, x_t = \sigma(t) d\, w_t \quad \mbox{where } \sigma(t) = \sigma_{\mathrm{min}} \( \frac{\sigma_{\mathrm{max}}}{\sigma_{\mathrm{min}}} \)^t
+#     d\, x_t = \sigma(t) d\, w_t \quad \mbox{where } \sigma(t) = \sigma_{\mathrm{min}}\left( \frac{\sigma_{\mathrm{max}}}{\sigma_{\mathrm{min}}}\right(^t
 
 from deepinv.sampling.sde import VESDE
 
@@ -99,7 +100,7 @@ dinv.utils.plot_videos(
 
 # %% Varying samples
 # One can obtain varying samples by using a different seed.
-# To ensure the reproducibility, if the parameter `rng` is given, the same sample will
+# To ensure the reproducibility, if the parameter :attrib:`rng` is given, the same sample will
 # be generated when the same seed is used
 solution = sde(
     (1, 3, 64, 64), timesteps=timesteps, method="Euler", seed=1, full_trajectory=True
@@ -116,7 +117,7 @@ dinv.utils.plot(solution.sample, suptitle=f"Backward sample, nfe = {solution.nfe
 # %% Plug-and-play Image Generation with arbitrary denoisers
 # ----------------------------------------------------------
 # The `SDE` class can be used together with any (well-trained) denoisers for image generation.
-# For example, we can use, for example the `DRUNet` for image generation.
+# For example, we can use, for example the :meth:`deepinv.models.DRUNet` for image generation.
 # However, it should be careful to set the parameter `rescale` to `True` when instantiating the SDE class, since
 # the DRUNet was trained on [0, 1] images.
 
@@ -172,8 +173,8 @@ solution = sde(
     (1, 3, 64, 64), timesteps=timesteps, method="Euler", seed=1, full_trajectory=True
 )
 
-# %% We could also use the DiffUNet.
-
+# %% We could also use the :meth:`deepinv.models.DRUNet`.
+#
 dinv.utils.plot_videos(
     solution.trajectory.cpu()[::4],
     time_dim=0,